Add createDialogStore for managing dialog state
This commit introduces a new file, createDialogStore.ts, for the creation of a dialog store. This store will be responsible for managing the state of dialog components. It includes methods for updating and resetting default values in the state. The creation of this store is necessary for a more efficient handling and control of dialog components within the application.