Export useCustomDialogStore directly
The useCustomDialogStore function is now being exported directly to align with the rest of our codebase. This was changed from default export to named export to improve the consistency of our exports across files resulting in a more predictable export structure. This change also involves the renaming of 'useCustomDialogStore.tsx' to 'useCustomDialogStore.ts' to signify that it doesn't render JSX.