Add useCategoryDialogStore.ts for category modifications
This commit introduces a new file, useCategoryDialogStore.ts, in the categories section of the store. It exports a function created using createDialogStore, with the default value for a category set up. This was done to enable easy modification of categories in the application. It uses the modify-category schema imported from ModifyCategoryDialog for validation.