Add AlbumDropDownActions component
A new component, AlbumDropDownActions, has been added to the AlbumsDataTable module. This component includes an Edit and Delete action for each album entry on the table. This is a piece of the larger task of refining the AlbumsDataTable UI, allowing for more specific actions to be performed directly from the table.