Add GET method to album routes and revalidate albums on POST
In this commit, the GET method was added to the album routes to fetch all albums. This change was necessary to support the retrieval of albums from the database. Additionally, the `revalidateTag` function was invoked after posting an album to revalidate and refresh the list of albums, ensuring that the latest information is always displayed.