"Added PUT and DELETE methods to announcement route"
In this commit, CRUD functionality has been expanded for the announcements route. Specifically, PUT and DELETE methods have been added and integrated with revalidation tags to ensure fresh data is always served. The changes will allow not just reading but also updating and deleting of announcements. The decision to enhance the CRUD functionality was taken to improve interactivity with announcement resources and provide more control over the data to the users.