Add SummaryShowFieldsCard component
This commit provides a new component named `SummaryShowFieldsCard`. This component will allow the user to see the summary of active and inactive fields in a table form within a card. This was added to improve the user's data visualization control and as a part of new feature implementation. An additional `summaryShowFieldCardTable.tsx` has been also added to support the main card component.