Add SummaryStatusCard component
Created a new SummaryStatusCard component and its corresponding SummaryStatusCardTable subcomponent. Both components are part of the new feature to display summary cards in the user interface. The SummaryStatusCard receives an array of records with statuses and displays counts of records with differing statuses in a table format in a card form. This change supports the UI enhancement to provide a visual summary of record status to users for easier tracking.