Add DataTableBooleanCell component
Added a new component, DataTableBooleanCell, to represent boolean information in tables. The component displays a colored cell with a text label. If the associated value is 'true', the cell's background will be green and display 'Tak'; if 'false', it will be red and display 'Nie'. This feature is important for visual segregation and quick readability of boolean data in tables.