Refactor Album page to use apiCall function
Switched from using getAlbumWithRelations to use apiCall function in the Admin Album page. This was done to ensure consistency and readability of the code. The apiCall function is used across the pages to retrieve data which adheres to our aim for encapsulated, reusable code. This includes changes to AlbumTabs and AlbumInfoCard to accommodate for the switch.