Refactor image upload and handling structure
Renamed and moved 'route-helper.ts' to a more centralized location in '/lib/services/images'. The code has been updated to use node:fs and node:path instead of the initially used path in order to align with the updated Node.js methodology.

The function 'checkDirectory' has been expanded to check for the existence of a specific album directory, instead of just the default upload directory. A new environment variable UPLOAD_DIR_ALBUM has been added to accommodate this change.

In 'validateImage', the ignored parameter in the destructuring operation is specifically expressed as '_key', to increase code readability.

Finally, a new function 'clearImage' has been added to allow for easy deletion of images from the server. This is intended to provide better data management and potentially faster processing times.