Refactor album directory creation to a separate service
The function to create a new directory for every new album was previously in the route file, which was cluttering the route file and not following the separation of concerns principle. This function has been moved to a new server-methods file under the services/albums directory. This makes the routing code more readable and maintainable, and allows reusability of the directory creation function.