Refactor AlbumEditContent to receive images as prop
This change moves the image data retrieval responsibility from the AlbumEditContent component itself to its parent component. Previously, AlbumEditContent was using the useUploadedImagesStore hook to access the images data. Now, the images are passed down as a prop. This increases the flexibility of the component and makes it more reusable.