Update AlbumCoverImage state management
Added a console.log for imageCoverId and a condition to setCoverImage from useAlbumCoverImageStore. This change aims to fix an issue where the coverImage was not updating properly when the imageCoverId was null. This will now ensure that the coverImage is correctly updated as per the imageCoverId, thereby improving data consistency.