Refactor menuContextContent component for better readability
The initial 'menuContextContent.tsx' file was removed and replaced with two individual files: 'deleteImageTrigger.tsx' and 'index.tsx'. This modification was done to improve the code maintainability and readability by splitting the large 'menuContextContent.tsx' into individual smaller components that each performs a single responsibility. The new structure also helps in isolating and identifying issues faster if any arises in the future.