Add Google and Facebook authentication
This commit introduces Google and Facebook as authentication providers. Additionally, a function has been added to check if a user's email is allowed access, by comparing it either to a list defined in the environment variables or to a list of administrators stored in the database. This ensures that only allowed users can login to the application.