Add administrator handling functions and components
This commit adds several new functions in the administrators' db lib allowing to get all administrators, get an administrator by email and delete an administrator by email. This functionality is provided to improve the management of the administrator users by offering the option to retrieve administrators' records and delete unnecessary records. Also, introduces new exported components in the administrators index, which are expected to be utilized for building administrators UI.