Add API call and components to Administrators Page
This commit includes the addition of an API call to fetch a list of administrators on the 'AdministratorsManagePage'. Also added the 'AdministratorsTable' and 'AddAdministratorDialog' components to the 'AdministratorsManagePage' to manage administrators more efficiently. This aids in enhancing the user interface and allows admins to manage other administrators from a single page.