Update package.json dependencies
This update adds new dependencies to the package.json file and updates the version of the existing Prisma client. The '@auth/prisma-adapter' and 'next-auth' are added to incorporate authentication for our app. 'nodemailer' is also added for handling emails. Furthermore, the Prisma client is updated to its latest version, 5.1.1, to take advantage of new features and fixes from the update.