Add administrator API calls
This commit introduces two new API calls related to the administrator's account management. The first one is used for creating a new administrator account, and the second one deletes an existing administrator account. This change will enable the front-end application to interact with the back-end system to manage administrator's accounts.