Add index file in album components
An index file was added in the 'album' components directory which now exports all the album-related components. This led to a change in imports in files where these components were used. The aim being to make the imports cleaner, more organized and easy to manage. This will also help to improve the readability of the code for future updates.