Consolidate component imports in login page
Reorganized the import statements in 'src/app/(auth)/login/page.tsx' for better readability by grouping them together and creating a new index file in 'src/components/login' to export all related components. Hereafter, all the components of the '/login' directory can be imported in a single line which makes the code cleaner and easier to read. This change doesn't affect any functionalities of the page.