Add customizable separator classes to CardHeaderWithSeparator
The `CardHeaderWithSeparator` component was extended to allow customization of the separator's classes. This change ensures that the styles of the separator can be controlled from the parent component, thereby increasing the reusability and flexibility of the `CardHeaderWithSeparator` component. The class names provided will be concatenated to the default 'w-11/12 mx-auto'.