Refactor DataTable to use Card component
The DataTableTitle component was removed and its functionality was replaced by Card and CardHeaderWithSeparator from the UI toolkit. This change leads to better consistency across the app and makes the DataTable more in-line with the existing UI/UX guidelines. The DataTable was also wrapped into the Card for better visual presentation. The separator logic was offloaded to the CardHeaderWithSeparator reducing clutter in the DataTable code.