Refactor useModifyAlbumLogic hook for album update and creation
This commit modifies the useModifyAlbumLogic function to properly update or create an album based on the passed id. Previously, the album CRUD operations for both update and create were performed regardless of the provided id. With this change, if an id is provided, the update operation is carried out, if not, a new album is created. This improves the logical flow and efficiency of the function.