Add "@radix-ui/react-scroll-area" package and update yarn.lock file
The new "@radix-ui/react-scroll-area" package has been added to package.json to provide improved scrollable area functionality to the application. The yarn.lock file has been updated to reflect this new addition along with changes in the other dependencies. The revisions in the yarn.lock file include updates to dependency versions and the addition of new dependencies required by the updated packages.