Move LogoutAlert component to alerts directory
The LogoutAlert component was moved from 'navigation/AdminNavigationSideBar' to 'alerts'. This change was made to maintain the proper organization of the project. The component is not restricted only to the AdminNavigationSideBar, hence it makes more sense to have it in a general 'alerts' directory. Additional changes include refining the component to receive 'className' as a prop for better component reusability.