Add ClientDesktopNavigationNestedList component
Added a new component, ClientDesktopNavigationNestedList, for desktop view. This component is responsible for rendering nested routes in the client's desktop navigation. It utilizes existing UI components such as NavigationMenuContent, NavigationMenuItem, NavigationMenuTrigger and a new subcomponent ListItem to display title and associated routes. This change is made to enhance navigation UX in desktop view.