Refactor MobileNavigationSheet and adjust styles
Updated the MobileNavigationSheet to accept a `renderItems` prop that allows any list of elements to be rendered dynamically. This change simplifies the component and improves its reusability as it does not depend on predefined navigationRoutes prop. Also removed a redundant `closeSheet` function in MobileNavigationSheet.

Also, various style adjustments have been made:
- Added "overflow-y-scroll" to SheetContent in MobileNavigationSheet to handle overflow content.
- Adjusted text size in mobileNavigationSheetHeader for better readability.
- Changed HamburgerMenuIcon color to black in mobileNavigationSheetTrigger for improved visibility.