"Add new NavigationMenu component"
This commit introduces a new NavigationMenu component to the UI library. Different parts of a navigational menu such as NavigationMenuList, NavigationMenuItem, NavigationMenuContent, NavigationMenuTrigger, NavigationMenuLink, NavigationMenuIndicator, and NavigationMenuViewport are defined and exported for use. These parts incorporate the @radix-ui/react-navigation-menu primitives. This component is essential to facilitate navigation through different parts of the application.