Refactor navigation components for better structure and clarity
The file names and component names have been updated for better understanding. "MobileNavigationDropdown" was renamed to "MobileListItem" and similarly "ClientDesktopNavigation" was renamed and moved into a more appropriate directory "ClientNavbar". This was done to convey a more accurate representation of what the components do. Additionally, arguments used in these components were simplified for better readability and maintainability. The changes have also included removing unneeded mentions of the "Logo" component in the "ClientDesktopNavigation", hence simplifying its structure.