Add Tooltip component and its dependency package
A new Tooltip component has been added to the project for providing interactive text hints to the user when they hover over or focus on an element. The component is built using the '@radix-ui/react-tooltip' package. This package has been added to the project's dependencies in the package.json file.