Add contact item copy functionality
This change adds a new feature to allow users to copy contact item information. A new component called `copyItem.tsx` has been added to the ContactItem directory which will display the copy text in a tooltip when clicked on. The copied text is saved to the user's clipboard. This makes it easier for users to quickly copy and use contact information.