Update ClientDesktopNavigationNestedList for balanced columns
Added a function calculateColSpan within the ClientDesktopNavigationNestedList component to evenly distribute the list items across the grid. Previously, if there was an odd number of navigation items, the last item would span one column and look awkward. The new function assigns two columns to the last item if the total number of items is odd, thus maintaining the seamless layout. In addition, destructuring and spread operator is used to simplify the mapping of nestedRoutes, improving the readability of the code.