Add WeekIntentionsTabContent component
A new functional component, WeekIntentionsTabContent, was added in the week-intentions directory. This component receives the day, the date of the day, intentions, and id as props. Some of the imported modules include TabsContent and weekdayTranslator. The component renders the day and date in the header and maps through the intentions to present them in the front-end. This was implemented to help better manage the weekly intentions presentation by condensing the related functionality into a single component.