Add ArticlePage component for individual articles
Created a new ArticlePage component to handle the display of individual articles. It uses utility functions to fetch article data based on the name parameter. If there is no corresponding article, it redirects user to a notFound page. This change is made to better structure our app and improve user navigation experience.

Signed-off-by: xyashino <szczypiropan123@gmial.com>