Add Bio page for St. Maksymilian Kolbe
A new page for the biography of St. Maksymilian Kolbe has been set up to provide information about his life and mission. This is done to educate the users about his life, decisions, and sacrifices for the faith. This page includes a header, biography text with various details about his life, and a source link for further reading. The decision to present this biography is due to St. Maksymilian Kolbe's significant contribution to the Christian faith.

Signed-off-by: xyashino <szczypiropan123@gmial.com>