Add Announcements Page
Introduced a new Announcement component in the frontend under src/app/(client)/announcements/page.tsx. This change was necessary to add the functionality of displaying active announcements on the website. The Announcement function fetches an active announcement and displays its details in an MDXRemote component with ImageTextSection and Card components for proper formatting and presentation.

Signed-off-by: xyashino <szczypiropan123@gmial.com>