Update layout styling in layout.tsx
Modified the CSS styling in layout.tsx to improve the UX design. The width of the main content div has been reduced slightly on large screens (from 11/12 to 10/12). Additional styling (rounded-xl) added to create rounded corners for a softer visual appearance. This change enhances the content readability and appearance for users.

Signed-off-by: xyashino <szczypiropan123@gmial.com>