Add Gallery page implementation
Gallery page was implemented in this commit. It renders CategorySections based on the category data fetched from the database using getCategoriesWithAlbums function. If no categories found, a redirect to home page is performed. This change was necessary to provide functionality for displaying albums under distinct categories in a user-friendly manner.

Signed-off-by: xyashino <szczypiropan123@gmial.com>