Refactor WeekIntentionContent and add intentions page
This commit removes unnecessary string interpolation for the 'id' within WeekIntentionContent component. Also, a new 'intentions' page is added to the app, which fetches 'weekIntentions' from a database and displays it. The motivation behind the 'intentions' page is to provide users with clear and updated information about the intentions for the week.

Signed-off-by: xyashino <szczypiropan123@gmial.com>