Update article type metadata details
The type ArticleImageMetadata in the src/types/article.d.ts file was commented out and a new type named SubArticleMetaData was added. The SubArticleMetaData is defined with two properties 'title' and 'order', and it's also combined with ImageMetadata.
This change was made to enhance the metadata structure and adapt to the new article model scheme.

Signed-off-by: xyashino <szczypiropan123@gmial.com>