Add office.mdx with Parish Office information
This commit adds a new file called office.mdx, which contains important information about the Parish Office. The document lists office hours along with the regulations on how, when, and to whom certain documents can be released. It specifically details the guidelines and rules associated with the EU's General Data Protection Regulation (GDPR) implemented in May 2018.

Signed-off-by: xyashino <szczypiropan123@gmial.com>