Add historical article about the parish in Gruszów Wielki
This commit introduces a new article into the articles directory - a historical overview of the parish in Gruszów Wielki. The article includes information dating back to 1405, lists of parish priests, and acknowledgments. It provides a historical context and valuable information to site visitors interested in the history of the parish.

Signed-off-by: xyashino <szczypiropan123@gmial.com>