Add funeral notice and procedure article
The new MDX file funeral.mdx has been created under src/articles. This is an informational article detailing the procedure for reporting a death, scheduling a funeral, and the documents necessary for a Catholic funeral. The aim is to guide the grieving families by making them aware of the procedures, requirements, and customary practices during such times. In addition, it also encourages parishioners to remember the departed in their prayers.

Signed-off-by: xyashino <szczypiropan123@gmial.com>