Add new Caritas article
A new article for the group Caritas has been added. This group is a parish branch of Caritas that provides help to those in need and a kitchen for the poor. The page includes information about the group, how to get involved, and how to contribute 1% of taxes to the group. This has been done to create a comprehensive source of information about the different groups on the website.

Signed-off-by: xyashino <szczypiropan123@gmial.com>