Add "Confirmation" article in the sacrament section
A new article on "Confirmation" has been added in the sacrament section for the website's religious content expansion. This document provides insights on the preparation, norms, and important details about receiving the sacrament of Confirmation in the parish. This way, it helps catechumens and their sponsors understand and prepare better for receiving this important rite.

Signed-off-by: xyashino <szczypiropan123@gmial.com>