Add new article about the sacrament of penance
A new article about the sacrament of penance was added to the 'sacrament' directory. This article provides important information about the sacrament itself, including when and how parishioners can receive it. It is an important resource for the community, particularly for those wanting to reconnect with their faith or learn more about the sacraments.

Signed-off-by: xyashino <szczypiropan123@gmial.com>